# Improve perf remove allocations